Create README.md
What can you do with your Slack data? Turns out, a lot! In this analysis, we go over the basics of how to build a network map in Python, in this instance, using your conversations in Slack regarding who is talking to who. From there, we dive deeper into the diagram by building a rule-based natural language processing model and a basic machine learning model to understand the context of the conversations that we've mapped. Which conversations are social and which are work-related? Which conversations are people asking for advice or technical questions? Who are the main players in the map who answer people's questions quickly and effectively? Through this process, we are able to find 10 'influencers' out of 200,000 Slack messages.